refactor: add filterRequestsByTitle controller
This commit adds a new controller function called filterRequestsByTitle. This function allows filtering requests by their title. It takes a title parameter and performs a database query to find requests that match the provided title. If any requests are found, it returns them along with the number of results. If no requests are found, it returns a 404 error. If there is an error during the query, it returns a 500 error.

Co-authored-by: [Author Name]